Don't subtract unsigned chars

+++ src/Utility/String.cpp
@@ -15,7 +15,7 @@ static int AsciiCaseInsensitiveCompare(const char *l, const char *r, size_t size
     const unsigned char *ur = reinterpret_cast<const unsigned char *>(r);
 
     for (;size > 0; size--) {
-        int result = AsciiToLower(*ul++) - AsciiToLower(*ur++);
+        int result = static_cast<int>(AsciiToLower(*ul++)) - static_cast<int>(AsciiToLower(*ur++));
         if (result != 0)
             return result;
     }
